Tweak Dev Env

+++ dev_scripts/setup.sh
@@ -15,30 +15,32 @@ fi
 #####################################################
 # Code below is to update the developer environment #
 #####################################################
+if [ -d "dev_vendor/php-cs-fixer" ]; then
 echo ""
 echo ""
 echo "Developer Environment Update Started"
 
 
 
-# Update WordPress ArrayDeclarationSpacingSniff if required
-WP_ArrayDeclarationSpacingSniff=dev_vendor/wp-coding-standards/wpcs/WordPress/Sniffs/Arrays/ArrayDeclarationSpacingSniff.php
-WP_AA_ArrayDeclarationSpacingSniff=dev_scripts/overrides/wpcs-ArrayDeclarationSpacingSniff.php
-WP_XX_ArrayDeclarationSpacingSniff=dev_vendor/wp-coding-standards/wpcs/WordPress/Sniffs/Arrays/updated
-if [ -e $WP_ArrayDeclarationSpacingSniff ] && [ -e $WP_AA_ArrayDeclarationSpacingSniff ] && [ ! -e $WP_XX_ArrayDeclarationSpacingSniff ];
-then
+	# Update WordPress ArrayDeclarationSpacingSniff if required
+	WP_ArrayDeclarationSpacingSniff=dev_vendor/wp-coding-standards/wpcs/WordPress/Sniffs/Arrays/ArrayDeclarationSpacingSniff.php
+	WP_AA_ArrayDeclarationSpacingSniff=dev_scripts/overrides/wpcs-ArrayDeclarationSpacingSniff.php
+	WP_XX_ArrayDeclarationSpacingSniff=dev_vendor/wp-coding-standards/wpcs/WordPress/Sniffs/Arrays/updated
+	if [ -e $WP_ArrayDeclarationSpacingSniff ] && [ -e $WP_AA_ArrayDeclarationSpacingSniff ] && [ ! -e $WP_XX_ArrayDeclarationSpacingSniff ];
+	then
 		echo "Overriding WordPress ArrayDeclarationSpacingSniff CS"
 		rm -f $WP_ArrayDeclarationSpacingSniff
 		cp $WP_AA_ArrayDeclarationSpacingSniff $WP_ArrayDeclarationSpacingSniff
 		touch $WP_XX_ArrayDeclarationSpacingSniff
-fi
+	fi
 
-# Set merge strategy in case of conflicts
-echo "Set merge strategy"
-git config merge.theirs.driver true
+	# Set strategy for merge conflicts
+	echo "Setting Merge Strategy"
+	git config merge.theirs.driver true
 
 
 
 echo "Developer Environment Update Ended"
 echo ""
 echo ""
+fi
